CA-4777 and CA-4778: Renaming order activity services
orders().Activities is now orders().activities()
OrdersActivitiesService is now OrderActivitiesService
